pyproject.toml
[project]
name = "pdftranslator"
version = "0.1.0"
description = "PDF translator with AI, preserving layout and images"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.7",
    "PyMuPDF>=1.24",
    "openai>=1.30",
    "anthropic>=0.30",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
]

[project.scripts]
pdftranslator = "pdftranslator.main:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
